requirements.txt
rich==13.7.0
python-dateutil==2.8.2
pytest==7.4.0